\name{fastOrdeR-package}
\alias{fastOrdeR-package}
\alias{fastOrdeR}
\docType{package}
\title{
\packageTitle{fastOrdeR}
}
\description{
\packageDescription{fastOrdeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastOrdeR}
\packageIndices{fastOrdeR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastOrdeR}

Maintainer: \packageMaintainer{fastOrdeR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
